+++ apps/web/hooks/queries/useDrinkListQuery.ts
@@ -9,11 +9,11 @@ import {
 
 const drinkListFetcher = async ({
   drinkType,
-  sorted = 'popularity_order',
+  sorted = 'popularity',
   page = 1,
 }: {
   drinkType: 'all' | 'beer' | 'tradition';
-  sorted?: 'popularity_order';
+  sorted?: 'popularity';
   page?: number;
 }) => {
   const response = await api
@@ -24,10 +24,10 @@ const drinkListFetcher = async ({
 
 export const useMainDrinkListQuery = ({
   drinkType,
-  sorted = 'popularity_order',
+  sorted = 'popularity',
 }: {
   drinkType: 'all' | 'beer' | 'tradition';
-  sorted?: 'popularity_order';
+  sorted?: 'popularity';
 }): UseQueryResult<DrinksResponse, Error> => {
   return useQuery({
     queryKey: ['mainDrinkList', drinkType, sorted],
@@ -37,10 +37,10 @@ export const useMainDrinkListQuery = ({
 
 export const useDrinkListQuery = ({
   drinkType,
-  sorted = 'popularity_order',
+  sorted = 'popularity',
 }: {
   drinkType: 'all' | 'beer' | 'tradition';
-  sorted?: 'popularity_order';
+  sorted?: 'popularity';
 }): UseInfiniteQueryResult<InfiniteData<DrinksResponse, Error>, Error> => {
   return useInfiniteQuery({
     queryKey: ['drinkList', drinkType, sorted],
