+++ packages/complete-ui/index.ts
@@ -1,7 +1,7 @@
 import './style.scss';
 
 import Counter from './src/Counter';
-import { Button, ChipButton } from './src/components/button/index';
+import { Button, ChipButton, CloseIconButton } from './src/components/button/index';
 import { Dropdown } from './src/components/dropdown';
 import { Input } from './src/components/input';
 import { TabList, Tab, Tabs, TabsPanel } from './src/components/tabs';
@@ -21,6 +21,7 @@ export {
   Counter,
   Button,
   ChipButton,
+  CloseIconButton,
   Dropdown,
   Input,
   TabList,
