+++ packages/complete-ui/src/components/button/index.ts
@@ -1,4 +1,5 @@
 import { StyledButton as Button } from '@/components/button/Button';
 import { ChipButton } from './chip/ChipButton';
+import CloseIconButton from './CloseButton';
 
-export { Button, ChipButton };
+export { Button, ChipButton, CloseIconButton };
