+++ packages/complete-ui/svg.d.ts
@@ -1,0 +1,4 @@
+declare module '*.svg' {
+    const content: React.FC<React.SVGProps<SVGElement>>
+    export default content
+}
