+++ packages/complete-ui/vite.config.ts
@@ -2,10 +2,11 @@ import { defineConfig } from 'vite';
 import { resolve } from 'path';
 import react from '@vitejs/plugin-react';
 import dts from 'vite-plugin-dts';
+import svgr from "vite-plugin-svgr";
 
 export default defineConfig({
-  plugins: [react(), dts({ copyDtsFiles: true })],
-  // assetsInclude: ['./src/assets**/*.jpg'],
+  plugins: [svgr(), react(), dts({ copyDtsFiles: true })],
+  assetsInclude: ['./src/assets**/*.svg'],
   resolve: {
     alias: {
       '@': '/src',
